We changed the cammands to a better name.
